updated the yml file to not include special chars
Changed strings like: abóbora_d'água to abobora_d_agua, since keys should
not contain special characters. Because of the change in the yml, I also
needed to change the csv and seed again.

While working on the yml I realised that the produce names appeared multiple
times, that's because I coppied from the csv, and there they appeared as many
time as they had months. So I cleaned that up in the yml file, now they are in
alphabetical order, and appear only once each.

Not related, I changed the factory_bot_rails version, it seems to be in 6.2.0