Feature – Find an produce item season by produce name (#30)
* feature/search_by_produce
This is the first version of finding by a produce name, it's very basic. We have a table with all produce in produce_items#index, we can see the months/season of the produce we chose in produce_items#show. It's easy to read, but we will still work on the style.

* Added formatted_month_name method to produce_items_helper
I thought we could use a helper to format months names, since that will probably appear in other places.

* Added a pt-BR.yml
We are using a dictionary(pt-BR.yml) to map the produce names, this way we don't need to use the readable_name method, we avoid logic and it's easier to translate in the future. 

To make it simpler to map, I changed the pt-BR.yml structure. It 'directly' maps our produce name to our produce_item names in the file. We are now able to write: I18n.t("produce_items.names.#{produce.name}")

Since all the produce is in portuguese, we are using pt-BR as default, and will make other translations as needed. We are translating everything to portuguese (eg.: categories, show and back links etc)

* Tests
Wrote models, controllers , views tests and e2e tests.

Added factory_bot and faker with their versions. This avoids unexpected surprises (like bugs caused by upgrading major versions without being aware)