first version of the item seasons index view
First I tried to write a test. I was thinking we would be displaying our
item seasons, and did the test with that in mind. But, we have many items
seasons, what we really are displaying are the 12 months.

We take our months from our dictionary, and get our month indexes from there.
I did a helper method to do that. We will need the indexes to conect with
our produce items.

I think that is probably the way to go, which means the view test doesn't
make sense anymore. I'll rethink it. And will make a helper test as well.