first version of item season show
The item season show view shows the month and a list of produce, it's name,
category and link to produce. I added the .include per bullet request to use
eager loading. I was making a N + 1 query.

I also added two actions to navigate to either the month list or produce list.
The simple back link could be confusing to the user at times.

I still need to work on view and integration tests.