trying some item seasons tests
I changed the request test, as it was failing. But it's a list that goes
trough our dictionary, I'm not sure if this test is worth keeping? Or if
we should be working with something like i18n-tasks. Here we are testing
i18n more than anything.

Regarding views, the index test is also pretty much testing i18n. I'm still
trying to figure out how to test show.