changed item season to be found by month index
It seemed to be working before, but it was confusing, and I'm not sure it
was really working properly. I changed it to use month_index as param, and
use that to find the season name and produce.

It still needs more care, but I think it's now going in the right direction.