transfered notice to application layouts
I removed the notice from all our individual views, and added it to the
application layout. This way we only have it in one place.

Besides that, I changed the <p> to an <aside>. I think it makes more sense
and is more semantic: The <aside> HTML element represents a portion of a
document whose content is only indirectly related to the document's main
content.