updated the logic to get the month names
We were getting an array of months from translations to then translate
back to an index. This also forced us to handle the fact that month_names
has nil as the first item.

Instead we are now getting the indices from a model to then derive a name
from it. We are also making use of our formatted_month_name helper.