added some tests for our root, rewrote with I18n
I added a controller test and a few integration tests for our root. And
rewrote it to use I18n.

Other than that: My linter suggested rewriting the unless in the
controller, so I rewrote it. I think it is still pretty readable, but
I'm not sure which is best.