added some styles to links and navbar
I also added some transitions to the rotation classes we will be using
on the months, and are currently using on the month and item link in
the home.

I added a the light-beige color to the body, because I noticed that in
come devices our home is short and we got a white color. I think it's
best to complete it with one of our colors.

Important: I wanted to add a line break to our logo in the YML file,
but it only worked if I call it like this t('root.title') instead of
this I18n.t('root.title') (with the I18n). Is there a reason we should
do one over the other?

Important: Rails decided to complain about me changing the images
directory name to img. So I changed it back, buuut it still isn't
finding the assets.