feat: main page/root and 404 (#34)
* Added the controller and routes for our main page

* Made the structure and tests for our main page.

* Deleted unused helpers

* Deleted the pages controller and views

At first I also built a controller and views for our error and extra 
pages. But after we discussed about the pages using slugs, 
I think it's best to keep it simple. From the start I should have only 
focused on the main page for this PR. So I removed this for now.

In another PR we will decide how to go about our errors, and if other 
extra pages are needed, we will discuss those separately.