reorganized the css a bit
renamed yellow-button to button--yellow to keep it a bit more
consistent. placed button-holder in a separeted file, as it is
a different component than button.

moved components/links to config/base, we are going to put
all the "classless" styles there.

I also moved the button sytles to the link, so that the whole
button is clickable.