changed font size and line height units
Out of habit I had used pt as units, but that
isn't great for screen readers. So I changed all
the font-size units to rems. I also made variables
named as xs, s, m, l, and xl, so it's easier to
apply the sizes.

I also changed the line heights to use proportions.
For the same reason as size was changed.