moved all the vars to base
I moved all the vars to root in base, I think it makes more
sense for all the vars to be in one place. And they all regard
the 'base' anyway.

I moved the text colors to the fonts file, and deleted the
colors file, since it wasn't needed anymore. And added my
base font to 'base' because it's the base font.

Changed white-txt to color-white as suggested by Tales.