worked on styles for item seasons index
.content: I took it out of the application view, and I'm using
it directly on each view. This way I can change the color, and
don't need to add an extra unnecessary div.

since I took it out of application, I also took it out of
application css, and created an own css for it.

index view: changed the table structure to a button one.
I'll need to update the tests.

month_short: since I'm using an abbreviated version of the
months names, created a new helper, and added those to i18n.

others: created background colors, going to need those since
it changes. created a size var based on nav, need it for the banners.