wip: restructuring table to cards
I first build everything as tables, but our layout is a card based one.
Other than that, I was trying to figure out the sorting, I tested
with the table in produce_items (that test is in this branch), but I'll
need to see how it will work for the cards. Probably in a different branch.