Updated tests & added a conditional month_index <= 12
Important 1:
we only want the 12 months to be displayed, if we got a month
that doesn't exist, say 52, it shouldn't break our site, and
it shouldn't be displayed.

Important 2:
'expect(helper.rotation(1 || 6))' was only checking for the 1.
It would never check the 6. So the whole test was re-written.

Besides that I cleaned up a bit:
the downcase in the helpers was not necessary, and the capitalize
in the view wasn't as well. Since we are displaying the months
in all caps trough the CSS. Because of this I also had to re-write
a test.