Feature/45/item seasons/index/frontend (#48)
* worked on styles for item seasons index

.content: I took it out of the application view, and I'm using
it directly on each view. This way I can change the color and
don't need to add an extra unnecessary div.

index view: updated the view's structure, and updated the tests

* month_short: since I'm using an abbreviated version of the
months names created a new helper and added those to i18n.

* created background colors, going to need those since
it changes. created a size var based on nav, need it for the banners.

* added rotation to the months' buttons using a helper, added a
test for the helper

* added height to HTML and body

I had added this while doing the main page, but during the
code review, I thought this was unnecessary. However, this is
needed so the footer will always stick to the bottom.

* imports from less specific to more specific

In order to enact proper cascading.
In this case, we'd start from config to components, and then to pages