added async to fontawesome script
We do that so the script won't block our page.

When the browser loads HTML and comes across a
<script>...</script> tag, it can't continue building the DOM.
It must execute the script right now.
That leads to two important issues:

  • Scripts can't see DOM elements below them,
  so they can't add handlers etc.
  • If there's a bulky script at the top of the page,
  it “blocks the page”. Users can't see the page content
  till it downloads and runs

Ways to solve this are: adding the script right before
the body closing tag, adding defer or async to the script.

Other than that: added some accessibility to our arrows.